requirements.txt
beautifulsoup4==4.4.1
DateTime==4.0.1
Flask==0.10.1
Flask-FlatPages==0.6
Frozen-Flask==0.12
google-api-python-client==1.4.2
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
livereload==2.4.1
lxml==3.5.0
Markdown==2.6.5
MarkupSafe==0.23
oauth2client==1.5.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pymongo==3.2.1
python-dateutil==2.4.2
python-docx==0.8.5
pytz==2015.7
PyYAML==3.11
rsa==3.3
simplejson==3.8.1
six==1.10.0
tornado==4.3
uritemplate==0.6
Werkzeug==0.11.3
zope.interface==4.1.3
